refactor(api): git merge 라우트를 전용 모듈로 추출
8개 라우트를 api/git/merge.py로 이동. 핸들러 본문·경로·함수명 무변경.
라우트 테이블 characterization 테스트 통과 — HTTP 표면 63건 동일, 가림 쌍 0건.

get_merge_service_for_project 헬퍼도 함께 옮겼다. 호출 8곳이 전부 merge
핸들러 안에 있어(실측) _shared.py 승격 대상이 아니다. 이 헬퍼가 빠지면서
_legacy.py 의 "Dependencies" 섹션이 통째로 비어 배너째 제거했다.

models.git 의 merge 계열 모델 8종과 can_merge_to_branch 도 미사용이 되어
_legacy.py 에서 제거했다 — 이제 그 이름들의 유일한 소비자는 merge.py 다.

_legacy.py 330줄. 잔여 라우트 14개는 Task 10(working_tree 11 + remote 3).
